repos : move github specific parts to own compilation unit
